feat(python): add iterrrows
