+++ py-polars/docs/source/reference/dataframe/modify_select.rst
@@ -27,6 +27,7 @@ Manipulation/selection
     DataFrame.insert_at_idx
     DataFrame.interpolate
     DataFrame.item
+    DataFrame.iterrows
     DataFrame.join
     DataFrame.join_asof
     DataFrame.limit
